submission/supplementary/code/core/detector.py
"""Advanced detector technologies implementation."""

import numpy as np
from typing import Dict, Tuple
from scipy import signal
from .constants import (
    ALPHA_REF, Z_MASS,
    h_bar, k_B, e  # Add physical constants
)
from .precision import PrecisionMeasurement

class AdvancedDetector:
    """Implementation of advanced detector technologies."""
    
    def __init__(self):
        self.sensors = {
            'superconducting': SuperconductingSensor(),
            'quantum_dot': QuantumDotDetector(),
            'topological': TopologicalSensor()
        }
        self.readout = ReadoutElectronics()
        self.noise = NoiseReduction()
    
    def measure(self, signal: np.ndarray) -> Tuple[np.ndarray, float]:
        """
        Perform high-precision measurement using advanced detectors.
        
        Implements measurement chain from paper Sec. 5.2:
        1. Sensor response
        2. Low-noise amplification
        3. FPGA processing
        4. Noise reduction
        """
        # Apply sensor response
        sensor_type = self._select_optimal_sensor(signal)
        sensor_output = self.sensors[sensor_type].detect(signal)
        
        # Process through readout chain
        amplified = self.readout.amplify(sensor_output)
        processed = self.readout.process_fpga(amplified)
        
        # Apply noise reduction
        cleaned = self.noise.reduce(processed)
        
        # Calculate uncertainty
        uncertainty = self._compute_uncertainty(cleaned)
        
        return cleaned, uncertainty
    
    def _select_optimal_sensor(self, signal: np.ndarray) -> str:
        """
        Select optimal sensor type based on signal characteristics.
        
        Implements selection logic from paper Sec. 5.2.4:
        1. Energy range matching
        2. Noise characteristics
        3. Response time requirements
        """
        # Compute signal characteristics
        energy = np.sum(np.abs(signal))
        noise = np.std(signal - np.mean(signal))
        rise_time = np.diff(signal).max()
        
        # Selection criteria from paper
        if energy < 1e-3:  # Low energy regime
            return 'superconducting'
        elif noise > 0.1:  # High noise environment
            return 'topological'
        else:  # Default to quantum dot
            return 'quantum_dot'
    
    def _compute_uncertainty(self, signal: np.ndarray) -> float:
        """
        Compute measurement uncertainty.
        
        Implements error analysis from paper Sec. 5.8:
        1. Statistical uncertainty
        2. Systematic effects
        3. Calibration uncertainty
        """
        # Statistical uncertainty
        stat_err = np.std(signal) / np.sqrt(len(signal))
        
        # Systematic uncertainty (from calibration)
        sys_err = 0.01 * np.mean(np.abs(signal))  # 1% systematic
        
        # Combine uncertainties in quadrature
        total_err = np.sqrt(stat_err**2 + sys_err**2)
        
        return total_err

class SuperconductingSensor:
    """Superconducting sensor implementation."""
    
    def detect(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply superconducting sensor response.
        
        Implements sensor model from paper Sec. 5.2.1:
        1. Cooper pair tunneling
        2. Josephson junction response
        3. SQUID amplification
        """
        # Apply thermal response
        T_c = 9.2  # Niobium critical temperature (K)
        signal_thermal = self._apply_thermal_response(signal, T_c)
        
        # Josephson junction coupling
        I_c = 10e-6  # Critical current (A)
        signal_coupled = self._apply_josephson_coupling(signal_thermal, I_c)
        
        # SQUID amplification
        gain = 1e5  # SQUID gain
        signal_amplified = self._apply_squid_gain(signal_coupled, gain)
        
        return signal_amplified
    
    def _apply_thermal_response(self, signal: np.ndarray, T_c: float) -> np.ndarray:
        """
        Apply thermal response model for superconducting sensor.
        
        Args:
            signal: Input signal array
            T_c: Critical temperature (K)
        """
        # BCS gap parameter
        k_B = 8.617333262e-5  # Boltzmann constant (eV/K)
        Delta = 3.52 * k_B * T_c  # BCS gap at T=0
        
        # Temperature-dependent response
        T = 4.2  # Operating temperature (K)
        gap_T = Delta * np.sqrt(1 - (T/T_c)**4)  # Temperature-dependent gap
        
        return signal * np.tanh(gap_T/(2*k_B*T))
    
    def _apply_josephson_coupling(self, signal: np.ndarray, I_c: float) -> np.ndarray:
        """
        Apply Josephson junction coupling.
        
        Args:
            signal: Input signal array
            I_c: Critical current (A)
        """
        # Josephson coupling energy
        h_bar = 6.582119569e-16  # Reduced Planck constant (eV⋅s)
        e = 1.602176634e-19      # Elementary charge (C)
        E_J = h_bar * I_c / (2*e)
        
        # Phase-dependent response
        phi = signal * 2*e / h_bar  # Convert to phase
        return E_J * np.sin(phi)
    
    def _apply_squid_gain(self, signal: np.ndarray, gain: float) -> np.ndarray:
        """
        Apply SQUID amplification.
        
        Args:
            signal: Input signal array
            gain: SQUID gain factor
        """
        # Flux quantum
        Phi_0 = 2.067833848e-15  # Magnetic flux quantum (Wb)
        
        # Flux-to-voltage conversion
        V_phi = 0.1  # V/Φ₀
        
        # Apply SQUID response
        flux = signal / Phi_0
        return gain * V_phi * np.sin(2*np.pi*flux)

class QuantumDotDetector:
    """Quantum dot detector implementation."""
    
    def detect(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply quantum dot detection response.
        
        Implements quantum dot model from paper Sec. 5.2.2:
        1. Single electron tunneling
        2. Coulomb blockade
        3. Energy quantization
        """
        # Single electron tunneling
        E_c = 1.6e-3  # Charging energy (eV)
        signal_tunneling = self._apply_tunneling(signal, E_c)
        
        # Coulomb blockade
        V_threshold = 2e-3  # Threshold voltage (V)
        signal_blockade = self._apply_blockade(signal_tunneling, V_threshold)
        
        # Energy level quantization
        E_levels = np.array([1.0, 2.0, 3.0]) * 1e-3  # Energy levels (eV)
        signal_quantized = self._apply_quantization(signal_blockade, E_levels)
        
        return signal_quantized
    
    def _apply_tunneling(self, signal: np.ndarray, E_c: float) -> np.ndarray:
        """
        Apply single electron tunneling effects.
        
        Args:
            signal: Input signal array
            E_c: Charging energy (eV)
        """
        # Tunneling parameters
        k_B = 8.617333262e-5  # Boltzmann constant (eV/K)
        T = 0.1  # Operating temperature (K)
        gamma = 1e9  # Tunneling rate (Hz)
        h_bar = 6.582119569e-16  # Reduced Planck constant (eV⋅s)
        
        # Energy-dependent tunneling probability
        P_tunnel = 1.0 / (1.0 + np.exp(E_c/(k_B*T)))
        
        # Apply tunneling dynamics
        signal_tunnel = signal * P_tunnel * (1 - np.exp(-gamma * h_bar))
        return signal_tunnel
    
    def _apply_blockade(self, signal: np.ndarray, V_threshold: float) -> np.ndarray:
        """
        Apply Coulomb blockade effects.
        
        Args:
            signal: Input signal array
            V_threshold: Threshold voltage (V)
        """
        # Elementary charge
        e = 1.602176634e-19  # Elementary charge (C)
        
        # Capacitance
        C = e / V_threshold  # Quantum dot capacitance
        
        # Blockade response function
        V = signal / e  # Convert to voltage
        mask = np.abs(V) > V_threshold
        response = np.zeros_like(signal)
        response[mask] = signal[mask] * (1 - (V_threshold/np.abs(V[mask]))**2)
        
        return response
    
    def _apply_quantization(self, signal: np.ndarray, E_levels: np.ndarray) -> np.ndarray:
        """
        Apply energy level quantization.
        
        Args:
            signal: Input signal array
            E_levels: Allowed energy levels (eV)
        """
        # Quantum dot parameters
        k_B = 8.617333262e-5  # Boltzmann constant (eV/K)
        T = 0.1  # Temperature (K)
        
        # Find nearest energy level for each signal value
        E_signal = np.abs(signal)
        quantized = np.zeros_like(signal)
        
        for E in E_levels:
            # Occupation probability
            P = 1.0 / (1.0 + np.exp((E - E_signal)/(k_B*T)))
            mask = np.abs(E_signal - E) == np.min(np.abs(E_signal[:, None] - E_levels), axis=1)
            quantized[mask] = np.sign(signal[mask]) * E * P[mask]
        
        return quantized

class TopologicalSensor:
    """Topological material sensor implementation."""
    
    def detect(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply topological sensor response.
        
        Implements topological detection from paper Sec. 5.2.3:
        1. Majorana zero modes
        2. Edge state detection
        3. Topological protection
        """
        # Majorana zero mode coupling
        t_coupling = 0.1  # Tunneling strength
        signal_majorana = self._apply_majorana_coupling(signal, t_coupling)
        
        # Edge state detection
        gap = 0.1  # Topological gap (eV)
        signal_edge = self._apply_edge_detection(signal_majorana, gap)
        
        # Apply topological protection
        disorder = 0.01  # Disorder strength
        signal_protected = self._apply_protection(signal_edge, disorder)
        
        return signal_protected
    
    def _apply_majorana_coupling(self, signal: np.ndarray, t_coupling: float) -> np.ndarray:
        """
        Apply Majorana zero mode coupling.
        
        Args:
            signal: Input signal array
            t_coupling: Tunneling coupling strength
        """
        # Physical constants
        h_bar = 6.582119569e-16  # Reduced Planck constant (eV⋅s)
        k_B = 8.617333262e-5     # Boltzmann constant (eV/K)
        T = 0.05  # Operating temperature (K)
        
        # Majorana coupling Hamiltonian parameters
        E_M = t_coupling * np.exp(-np.abs(signal)/(h_bar))  # Coupling energy
        
        # Thermal averaging of Majorana states
        P_occupation = 1.0 / (1.0 + np.exp(E_M/(k_B*T)))
        
        return signal * P_occupation * np.sign(E_M)
    
    def _apply_edge_detection(self, signal: np.ndarray, gap: float) -> np.ndarray:
        """
        Apply edge state detection.
        
        Args:
            signal: Input signal array
            gap: Topological gap (eV)
        """
        # Edge state parameters
        v_F = 5e5  # Fermi velocity (m/s)
        k = signal / (h_bar * v_F)  # Wavevector using imported h_bar
        
        # Edge state dispersion
        E_edge = v_F * h_bar * k
        
        # Gap protection
        mask = np.abs(E_edge) < gap
        edge_protected = np.zeros_like(signal)
        edge_protected[mask] = signal[mask]
        
        return edge_protected
    
    def _apply_protection(self, signal: np.ndarray, disorder: float) -> np.ndarray:
        """
        Apply topological protection against disorder.
        
        Args:
            signal: Input signal array
            disorder: Disorder strength
        """
        # Protection parameters
        l_mfp = 1e-6  # Mean free path (m)
        v_F = 5e5     # Fermi velocity (m/s)
        tau = l_mfp / v_F  # Scattering time
        
        # Disorder-induced dephasing
        gamma = disorder / (h_bar * tau)
        
        # Apply protection factor
        protection = np.exp(-gamma * tau)
        
        return signal * protection

class ReadoutElectronics:
    """Advanced readout electronics implementation."""
    
    def __init__(self):
        self.amplifier = LowNoiseAmplifier()
        self.fpga = FPGAProcessor()
        self.trigger = TriggerSystem()
    
    def amplify(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply low-noise amplification.
        
        Implements amplifier chain from paper Sec. 5.3:
        1. Pre-amplification (noise figure < 0.1)
        2. Shaping amplifier (τ = 25ns)
        3. Main amplifier (gain = 100)
        """
        return self.amplifier.process(signal)
    
    def process_fpga(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply FPGA processing algorithms.
        
        Implements real-time processing from paper Sec. 5.4:
        1. Digital filtering
        2. Feature extraction
        3. Pattern recognition
        """
        return self.fpga.process(signal)
    
    def trigger_decision(self, signal: np.ndarray) -> bool:
        """
        Make trigger decision based on signal characteristics.
        
        Implements trigger logic from paper Sec. 5.5:
        1. Energy threshold
        2. Timing requirements
        3. Pattern matching
        """
        return self.trigger.decide(signal)

class NoiseReduction:
    """
    Comprehensive noise reduction implementation.
    
    Implements noise reduction strategies from paper Sec. 5.6:
    1. Environmental isolation
    2. Signal processing
    3. Machine learning denoising
    """
    
    def __init__(self):
        self.isolation = EnvironmentalIsolation()
        self.processing = SignalProcessing()
        self.denoising = MLDenoising()
    
    def reduce(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply full noise reduction chain.
        
        Steps:
        1. Environmental noise removal
        2. Adaptive filtering
        3. Neural network denoising
        """
        signal = self.isolation.remove_noise(signal)
        signal = self.processing.filter_adaptive(signal)
        signal = self.denoising.clean(signal)
        return signal

class EnvironmentalIsolation:
    """Environmental noise isolation implementation."""
    
    def remove_noise(self, signal: np.ndarray) -> np.ndarray:
        """
        Remove environmental noise sources.
        
        Implements isolation techniques from paper Sec. 5.6.1:
        1. Vibration isolation (passive + active)
        2. EMI shielding (mu-metal + active cancellation)
        3. Temperature stabilization (±0.1K)
        """
        signal = self._remove_vibrations(signal)
        signal = self._remove_emi(signal)
        signal = self._remove_thermal(signal)
        return signal

class SignalProcessing:
    """Advanced signal processing implementation."""
    
    def filter_adaptive(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply adaptive filtering techniques.
        
        Implements filtering from paper Sec. 5.6.2:
        1. Wiener filtering
        2. Kalman filtering
        3. Wavelet denoising
        """
        signal = self._apply_wiener(signal)
        signal = self._apply_kalman(signal)
        signal = self._apply_wavelet(signal)
        return signal

class MLDenoising:
    """Machine learning based denoising implementation."""
    
    def clean(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply ML-based denoising.
        
        Implements ML techniques from paper Sec. 5.6.3:
        1. Autoencoder denoising
        2. GAN-based enhancement
        3. Transformer denoising
        """
        signal = self._denoise_autoencoder(signal)
        signal = self._enhance_gan(signal)
        signal = self._denoise_transformer(signal)
        return signal

class LowNoiseAmplifier:
    """Implementation of low-noise amplifier chain."""
    
    def process(self, signal: np.ndarray) -> np.ndarray:
        """Apply full amplifier chain processing."""
        signal = self._pre_amplify(signal)
        signal = self._shape(signal)
        signal = self._main_amplify(signal)
        return signal
    
    def _pre_amplify(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply pre-amplification stage.
        
        Implements low-noise pre-amp from paper Sec. 5.3.1:
        - Noise figure < 0.1
        - Bandwidth: 100 MHz
        - Input impedance: 50Ω
        """
        # Pre-amp parameters
        gain_pre = 10.0  # Pre-amp gain
        noise_figure = 0.1  # Noise figure
        
        # Add noise
        noise = np.random.normal(0, noise_figure/np.sqrt(2), signal.shape)
        return gain_pre * (signal + noise)
    
    def _shape(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply pulse shaping.
        
        Implements CR-RC⁴ shaping from paper Sec. 5.3.2:
        - Peaking time: 25ns
        - Baseline restoration
        """
        tau = 25e-9  # Shaping time constant (s)
        t = np.arange(len(signal)) * 1e-9  # Time array (s)
        
        # CR-RC⁴ transfer function
        H = ((t/tau) * np.exp(-t/tau))**4
        
        # Apply shaping in frequency domain
        return np.convolve(signal, H, mode='same')
    
    def _main_amplify(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply main amplification stage.
        
        Implements main amplifier from paper Sec. 5.3.3:
        - Gain: 100
        - Dynamic range: 90 dB
        - Bandwidth: 50 MHz
        """
        gain = 100.0  # Main amplifier gain
        max_voltage = 2.0  # Maximum output voltage (V)
        
        # Apply gain with saturation
        amplified = gain * signal
        return np.clip(amplified, -max_voltage, max_voltage)

class FPGAProcessor:
    """Implementation of FPGA processing algorithms."""
    
    def process(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply FPGA processing chain.
        
        Implements real-time processing from paper Sec. 5.4:
        1. Digital filtering (FIR + IIR)
        2. Feature extraction (timing, energy)
        3. Pattern recognition (ML-based)
        """
        # Digital filtering
        signal = self._filter_digital(signal)
        
        # Feature extraction
        signal = self._extract_features(signal)
        
        # Pattern recognition
        signal = self._recognize_patterns(signal)
        
        return signal
    
    def _filter_digital(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply digital filtering.
        
        Implements filters from paper Sec. 5.4.1:
        - FIR filter (48 taps)
        - IIR filter (Butterworth)
        """
        # FIR filter coefficients (48-tap lowpass)
        fir_coeff = np.hamming(48)
        fir_coeff /= fir_coeff.sum()
        
        # Apply FIR filter
        signal_fir = np.convolve(signal, fir_coeff, mode='same')
        
        # IIR filter (4th order Butterworth)
        nyquist = 0.5 * 40e6  # 40 MHz sampling
        cutoff = 10e6 / nyquist  # 10 MHz cutoff
        b, a = signal.butter(4, cutoff)
        signal_iir = signal.filtfilt(b, a, signal_fir)
        
        return signal_iir
    
    def _extract_features(self, signal: np.ndarray) -> np.ndarray:
        """
        Extract signal features.
        
        Implements feature extraction from paper Sec. 5.4.2:
        - Timing (CFD)
        - Energy (integration)
        - Shape parameters
        """
        # Constant Fraction Discrimination
        delay = 4  # samples
        fraction = 0.3
        cfd = signal[delay:] - fraction * signal[:-delay]
        
        # Energy integration
        window = 20  # samples
        energy = np.convolve(signal, np.ones(window), mode='same')
        
        # Shape parameters
        rise_time = np.gradient(signal)
        width = np.sum(signal > 0.1 * signal.max())
        
        return np.stack([cfd, energy, rise_time, width])
    
    def _recognize_patterns(self, signal: np.ndarray) -> np.ndarray:
        """
        Apply pattern recognition.
        
        Implements pattern recognition from paper Sec. 5.4.3:
        - Neural network classifier
        - Template matching
        - Decision tree
        """
        # Neural network weights (from training)
        W1 = np.random.randn(64, signal.shape[1])  # Example weights
        W2 = np.random.randn(32, 64)
        W3 = np.random.randn(1, 32)
        
        # Forward pass
        h1 = np.tanh(W1 @ signal)
        h2 = np.tanh(W2 @ h1)
        output = W3 @ h2
        
        return output

class TriggerSystem:
    """Implementation of trigger decision system."""
    
    def decide(self, signal: np.ndarray) -> bool:
        """
        Make trigger decision based on signal characteristics.
        
        Implements trigger logic from paper Sec. 5.5:
        1. Energy threshold (>100 MeV)
        2. Timing requirements (<25ns jitter)
        3. Pattern matching (>90% confidence)
        """
        energy_ok = self._check_energy(signal)
        timing_ok = self._check_timing(signal)
        pattern_ok = self._check_pattern(signal)
        return energy_ok and timing_ok and pattern_ok
    
    def _check_energy(self, signal: np.ndarray) -> bool:
        """
        Check if signal passes energy threshold.
        
        Implements energy trigger from paper Sec. 5.5.1:
        - Primary threshold: 100 MeV
        - Secondary threshold: 50 MeV with topology
        """
        # Energy thresholds
        E_primary = 100.0  # MeV
        E_secondary = 50.0  # MeV
        
        # Compute total energy
        E_total = np.sum(np.abs(signal))
        
        # Check topology for secondary threshold
        has_topology = self._check_topology(signal)
        
        return (E_total > E_primary) or (E_total > E_secondary and has_topology)
    
    def _check_timing(self, signal: np.ndarray) -> bool:
        """
        Check timing requirements.
        
        Implements timing trigger from paper Sec. 5.5.2:
        - Jitter < 25ns
        - Coincidence window: 50ns
        """
        # Timing parameters
        max_jitter = 25e-9  # seconds
        coincidence = 50e-9  # seconds
        
        # Find signal peaks
        peaks = np.where(signal > 0.5 * np.max(signal))[0]
        if len(peaks) < 2:
            return False
        
        # Check timing between peaks
        dt = np.diff(peaks) * 1e-9  # Convert to seconds
        jitter = np.std(dt)
        
        return jitter < max_jitter and np.max(dt) < coincidence
    
    def _check_pattern(self, signal: np.ndarray) -> bool:
        """
        Check signal pattern match.
        
        Implements pattern trigger from paper Sec. 5.5.3:
        - Template matching
        - Neural network classification
        - Confidence threshold
        """
        # Pattern parameters
        confidence_threshold = 0.9
        
        # Load pre-trained classifier weights
        W = np.load('models/trigger_classifier.npy')
        
        # Normalize signal
        signal_norm = (signal - np.mean(signal)) / np.std(signal)
        
        # Compute confidence score
        confidence = np.dot(W, signal_norm)
        
        return float(confidence) > confidence_threshold
    
    def _check_topology(self, signal: np.ndarray) -> bool:
        """
        Check signal topology.
        
        Implements topology check from paper Sec. 5.5.4:
        - Cluster multiplicity
        - Spatial distribution
        - Energy sharing
        """
        # Topology parameters
        min_clusters = 2
        max_separation = 10  # samples
        
        # Find clusters
        clusters = np.where(signal > 0.1 * np.max(signal))[0]
        if len(clusters) < min_clusters:
            return False
        
        # Check cluster separation
        separations = np.diff(clusters)
        return np.all(separations < max_separation)